move sampler AgateEnvironment to ThemeEnvironment
