--- samples/sampler/src/AgateEnvironment/package.json
@@ -1,3 +1,0 @@
-{
-	"main": "AgateEnvironment.js"
-}
