+++ samples/sampler/src/ThemeEnvironment/ThemeEnvironment.module.less
@@ -1,4 +1,4 @@
-// AgateEnvironment.module.less
+// ThemeEnvironment.module.less
 //
 @import "~../../../../styles/mixins.less";
 @import "~../../../../styles/skin.less";
